+++ src/shared/components/routes.js
@@ -2,13 +2,13 @@ import React from 'react';
 import {Route, IndexRoute} from 'react-router';
 import AppContainer from '../containers/app-container';
 import IndexContainer from '../containers/index-container';
-import MapLayout from '../components/MapLayout';
+import MapContainer from '../containers/MapContainer';
 import ListLayout from '../components/ListLayout';
 
 export default (
   <Route path="/" component={AppContainer}>
     <IndexRoute component={IndexContainer}/>
-    <Route path="map" component={MapLayout}/>
+    <Route path="map" component={MapContainer}/>
     <Route path="list" component={ListLayout}/>
   </Route>
 );
